entrypoint.sh: Modify to use "$@"
Modify entrypoint.sh, used by the Dockerfile.

The previous version of entrypoint.sh only used the first parameter.
What's worse, it would double-expand globs
(referring to "*.c", if a file was named *.c, that would get
expanded again). Usually variable references in shell are quoted.

In addition, the original version *ALWAYS* echoed a success,
even if the command did NOT succeed for some reason.
Instead of printing the spurious message, just show the output and
let the exit value get communicated back to the caller.
This is especially important for CI/CD, since we want the CI/CD
system to get the exit value (e.g., so it can report failure if there
was a failure).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>