entrypoint.sh: Small improvements
We can't quote argument 1, as it is expected to contain
multiple whitespace-seprated parameters and we need to expand any
globs in it.

This adds the use of "tee", so that the output goes to both the
file specified in $2 and also to stdout (so the results are directly
visible in the CI/CD display).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>